[Security Solution] - move network flyout to expandable framework
