+++ x-pack/plugins/security_solution/public/flyout/network_details/components/network_details.tsx
@@ -1,0 +1,118 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0; you may not use this file except in compliance with the Elastic License
+ * 2.0.
+ */
+
+import React, { useCallback, useMemo } from 'react';
+import { useDispatch } from 'react-redux';
+import { getEsQueryConfig } from '@kbn/data-plugin/common';
+import { InputsModelId } from '../../../common/store/inputs/constants';
+import { useInvalidFilterQuery } from '../../../common/hooks/use_invalid_filter_query';
+import type { FlowTargetSourceDest } from '../../../../common/search_strategy';
+import { IpOverview } from '../../../explore/network/components/details';
+import { useDeepEqualSelector } from '../../../common/hooks/use_selector';
+import { useGlobalTime } from '../../../common/containers/use_global_time';
+import { networkToCriteria } from '../../../common/components/ml/criteria/network_to_criteria';
+import { scoreIntervalToDateTime } from '../../../common/components/ml/score/score_interval_to_datetime';
+import { useKibana } from '../../../common/lib/kibana';
+import { convertToBuildEsQuery } from '../../../common/lib/kuery';
+import { inputsSelectors } from '../../../common/store';
+import { setAbsoluteRangeDatePicker } from '../../../common/store/inputs/actions';
+import { useSourcererDataView } from '../../../sourcerer/containers';
+import { useNetworkDetails } from '../../../explore/network/containers/details';
+import { networkModel } from '../../../explore/network/store';
+import { useAnomaliesTableData } from '../../../common/components/ml/anomaly/use_anomalies_table_data';
+import { useInstalledSecurityJobNameById } from '../../../common/components/ml/hooks/use_installed_security_jobs';
+import { EmptyPrompt } from '../../../common/components/empty_prompt';
+
+export interface NetworkDetailsProps {
+  expandedNetwork: { ip: string; flowTarget: FlowTargetSourceDest };
+}
+
+export const NetworkDetails = ({
+  contextID,
+  expandedNetwork,
+  isDraggable,
+}: NetworkDetailsProps & { contextID: string; isDraggable?: boolean }) => {
+  const { ip, flowTarget } = expandedNetwork;
+  const dispatch = useDispatch();
+  const { to, from, isInitializing } = useGlobalTime();
+  const getGlobalQuerySelector = useMemo(() => inputsSelectors.globalQuerySelector(), []);
+  const getGlobalFiltersQuerySelector = useMemo(
+    () => inputsSelectors.globalFiltersQuerySelector(),
+    []
+  );
+
+  const query = useDeepEqualSelector(getGlobalQuerySelector);
+  const filters = useDeepEqualSelector(getGlobalFiltersQuerySelector);
+
+  const type = networkModel.NetworkType.details;
+  const narrowDateRange = useCallback(
+    (score, interval) => {
+      const fromTo = scoreIntervalToDateTime(score, interval);
+      dispatch(
+        setAbsoluteRangeDatePicker({
+          id: InputsModelId.global,
+          from: fromTo.from,
+          to: fromTo.to,
+        })
+      );
+    },
+    [dispatch]
+  );
+  const {
+    services: { uiSettings },
+  } = useKibana();
+
+  const { indicesExist, indexPattern, selectedPatterns } = useSourcererDataView();
+  const [filterQuery, kqlError] = convertToBuildEsQuery({
+    config: getEsQueryConfig(uiSettings),
+    indexPattern,
+    queries: [query],
+    filters,
+  });
+
+  const [loading, { id, networkDetails }] = useNetworkDetails({
+    skip: isInitializing || filterQuery === undefined,
+    filterQuery,
+    indexNames: selectedPatterns,
+    ip,
+  });
+
+  useInvalidFilterQuery({ id, filterQuery, kqlError, query, startDate: from, endDate: to });
+  const { jobNameById } = useInstalledSecurityJobNameById();
+  const jobIds = useMemo(() => Object.keys(jobNameById), [jobNameById]);
+  const [isLoadingAnomaliesData, anomaliesData] = useAnomaliesTableData({
+    criteriaFields: networkToCriteria(ip, flowTarget),
+    startDate: from,
+    endDate: to,
+    skip: isInitializing,
+    jobIds,
+    aggregationInterval: 'auto',
+  });
+
+  return indicesExist ? (
+    <IpOverview
+      contextID={contextID}
+      id={id}
+      ip={ip}
+      data={networkDetails}
+      anomaliesData={anomaliesData}
+      loading={loading}
+      isInDetailsSidePanel
+      isLoadingAnomaliesData={isLoadingAnomaliesData}
+      isDraggable={isDraggable}
+      type={type}
+      flowTarget={flowTarget}
+      startDate={from}
+      endDate={to}
+      narrowDateRange={narrowDateRange}
+      indexPatterns={selectedPatterns}
+      jobNameById={jobNameById}
+    />
+  ) : (
+    <EmptyPrompt />
+  );
+};
