+++ x-pack/plugins/security_solution/public/timelines/components/formatted_ip/index.test.tsx
@@ -13,6 +13,7 @@ import { TestProviders } from '../../../common/mock';
 import { TimelineId, TimelineTabs } from '../../../../common/types/timeline';
 import { timelineActions } from '../../store';
 import { StatefulEventContext } from '../../../common/components/events_viewer/stateful_event_context';
+import { NetworkPanelKey } from '../../../flyout/network_details';
 
 jest.mock('react-redux', () => {
   const origin = jest.requireActual('react-redux');
@@ -53,6 +54,13 @@ jest.mock('../../store', () => {
     },
   };
 });
+
+const mockOpenFlyout = jest.fn();
+jest.mock('@kbn/expandable-flyout', () => ({
+  useExpandableFlyoutApi: () => ({
+    openFlyout: mockOpenFlyout,
+  }),
+}));
 
 describe('FormattedIp', () => {
   const props = {
@@ -100,7 +108,7 @@ describe('FormattedIp', () => {
     expect(timelineActions.toggleDetailPanel).not.toHaveBeenCalled();
   });
 
-  test('if enableIpDetailsFlyout, should open NetworkDetailsSidePanel', () => {
+  test('if enableIpDetailsFlyout, should open NetworkDetails expandable flyout', () => {
     const context = {
       enableHostDetailsFlyout: true,
       enableIpDetailsFlyout: true,
@@ -116,14 +124,17 @@ describe('FormattedIp', () => {
     );
 
     userEvent.click(screen.getByTestId('network-details'));
-    expect(timelineActions.toggleDetailPanel).toHaveBeenCalledWith({
-      id: context.timelineID,
-      panelView: 'networkDetail',
-      params: {
-        flowTarget: 'source',
-        ip: props.value,
+    expect(mockOpenFlyout).toHaveBeenCalledWith({
+      right: {
+        id: NetworkPanelKey,
+        params: {
+          contextId: 'timeline-1',
+          expandedNetwork: {
+            ip: props.value,
+            flowTarget: 'source',
+          },
+        },
       },
-      tabType: context.tabType,
     });
   });
 });
