+++ x-pack/plugins/security_solution/public/timelines/components/formatted_ip/index.tsx
@@ -7,13 +7,11 @@
 
 import { isArray, isEmpty, isString, uniq } from 'lodash/fp';
 import React, { useCallback, useMemo, useContext } from 'react';
-import { useDispatch } from 'react-redux';
 import deepEqual from 'fast-deep-equal';
 
 import type { EuiButtonEmpty, EuiButtonIcon } from '@elastic/eui';
-import type { ExpandedDetailType } from '../../../../common/types';
+import { useExpandableFlyoutApi } from '@kbn/expandable-flyout';
 import { StatefulEventContext } from '../../../common/components/events_viewer/stateful_event_context';
-import { getScopedActions } from '../../../helpers';
 import { FlowTargetSourceDest } from '../../../../common/search_strategy/security_solution/network';
 import {
   DragEffects,
@@ -26,8 +24,8 @@ import { parseQueryValue } from '../timeline/body/renderers/parse_query_value';
 import type { DataProvider } from '../timeline/data_providers/data_provider';
 import { IS_OPERATOR } from '../timeline/data_providers/data_provider';
 import { Provider } from '../timeline/data_providers/provider';
-import type { TimelineTabs } from '../../../../common/types/timeline';
 import { NetworkDetailsLink } from '../../../common/components/links';
+import { NetworkPanelKey } from '../../../flyout/network_details';
 
 const getUniqueId = ({
   contextId,
@@ -165,6 +163,8 @@ const AddressLinksItemComponent: React.FC<AddressLinksItemProps> = ({
   truncate,
   title,
 }) => {
+  const { openFlyout } = useExpandableFlyoutApi();
+
   const key = `address-links-draggable-wrapper-${getUniqueId({
     contextId,
     eventId,
@@ -177,7 +177,6 @@ const AddressLinksItemComponent: React.FC<AddressLinksItemProps> = ({
     [address, contextId, eventId, fieldName]
   );
 
-  const dispatch = useDispatch();
   const eventContext = useContext(StatefulEventContext);
   const isInTimelineContext =
     address && eventContext?.enableIpDetailsFlyout && eventContext?.timelineID;
@@ -190,29 +189,24 @@ const AddressLinksItemComponent: React.FC<AddressLinksItemProps> = ({
       }
 
       if (eventContext && isInTimelineContext) {
-        const { tabType, timelineID } = eventContext;
-        const updatedExpandedDetail: ExpandedDetailType = {
-          panelView: 'networkDetail',
-          params: {
-            ip: address,
-            flowTarget: fieldName.includes(FlowTargetSourceDest.destination)
-              ? FlowTargetSourceDest.destination
-              : FlowTargetSourceDest.source,
+        const { timelineID } = eventContext;
+        openFlyout({
+          right: {
+            id: NetworkPanelKey,
+            params: {
+              contextId: timelineID,
+              expandedNetwork: {
+                ip: address,
+                flowTarget: fieldName.includes(FlowTargetSourceDest.destination)
+                  ? FlowTargetSourceDest.destination
+                  : FlowTargetSourceDest.source,
+              },
+            },
           },
-        };
-        const scopedActions = getScopedActions(timelineID);
-        if (scopedActions) {
-          dispatch(
-            scopedActions.toggleDetailPanel({
-              ...updatedExpandedDetail,
-              id: timelineID,
-              tabType: tabType as TimelineTabs,
-            })
-          );
-        }
+        });
       }
     },
-    [onClick, eventContext, isInTimelineContext, address, fieldName, dispatch]
+    [onClick, eventContext, isInTimelineContext, address, fieldName, openFlyout]
   );
 
   // The below is explicitly defined this way as the onClick takes precedence when it and the href are both defined
